refactoring raspberry machine h and cc, fixing raspberry ic asserts comparison, fixing raspberry uart reference to loopback method, add verification to only wakeup secondary cores if traits is configured to more than 1 cpu. Tried to compile, got an error on ic mapping of mailbox eoi on raspberry_pi3_ic_init.cc

+++ include/machine/cortex/raspberry_pi3/raspberry_pi3_ic.h
@@ -40,7 +40,7 @@ public:
     }
 
     static void enable(int i) {
-        assert(i <= INTS);
+        assert((unsigned int)i <= INTS);
         switch(i / 32){
         case 0 /* */:
         case 1 /* */:
@@ -54,7 +54,7 @@ public:
     }
 
     static void disable(int i) {
-        assert(i <= INTS);
+        assert((unsigned int)i <= INTS);
         switch(i / 32){
         case 0 /* */:
         case 1 /* */:
